refactored accordion for allies and enemies with comments
