+++ src/Avatar/pages/AvatarsPage.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from 'react';
 import { ErrorBoundary, useErrorHandler } from 'react-error-boundary';
 import { CharacterCard } from '../../ui-core/CharacterCard';
 import { ErrorFallback } from '../../ui-core/ErrorBoundary';
-import { AvatarType } from '../avatars';
 import { lastAirBenderApi } from '../constants';
+import { AvatarType } from '../type';
 
 const AvatarsPage = () => {
   const [avatars, setAvatars] = useState<Array<AvatarType>>([]);
