+++ src/Avatar/type.ts
